+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix get_ticket_number when branch name contains numbers
 - Load `.env.local` on top of `.env` in case it exists
+- Add `{{BACKGROUND}}` with "Details in the ticket" by default when a ticket link is used
 
 ## 0.2 - 2024-08-28